finish chapter 5

+++ src/items/items.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { User } from 'src/entities/user.entity';
 import { Item } from '../entities/item.entity';
 import { CreateItemDto } from './dto/create-item.dto';
 import { ItemStatus } from './item-status.enum';
@@ -18,14 +23,17 @@ export class ItemsService {
     return await this.itemRepository.findOne(id);
   }
 
-  async create(createItemDto: CreateItemDto): Promise<Item> {
-    return await this.itemRepository.createItem(createItemDto);
+  async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
+    return await this.itemRepository.createItem(createItemDto, user);
   }
 
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.findById(id);
     if (!item) {
       throw new NotFoundException();
+    }
+    if (item.userId === user.id) {
+      throw new BadRequestException('自分の商品を購入することはできません');
     }
     const newItem = {
       ...item,
@@ -37,7 +45,14 @@ export class ItemsService {
     return newItem;
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
+    const item = await this.findById(id);
+    if (!item) {
+      throw new NotFoundException();
+    }
+    if (item.userId !== user.id) {
+      throw new BadRequestException('自分の商品以外は削除できません');
+    }
     await this.itemRepository.delete(id);
   }
 }
